Aarch64 Push and pop xzr for irrelevant values
Summary:
The Aarch64 stack frame must be aligned to 16 bytes, so push and
pop the xzr register for irrelevant values. The xzr can be slightly
faster than rAsm or duplicate LRs on some cpus. It can also make
the stack easier to debug as it will always be read as zeroes
instead of duplicate or junk values.